Removed BuilderApiState trait to complete simplifying abstraction and consolidated mev_boost module into routes module. Also stop using hardcoded ports to prevent clashing
