+++ bin/src/lib.rs
@@ -17,10 +17,7 @@ pub mod prelude {
         utils::{initialize_tracing_log, utcnow_ms, utcnow_ns, utcnow_sec, utcnow_us},
     };
     pub use cb_metrics::provider::MetricsProvider;
-    pub use cb_pbs::{
-        BuilderApiState, PbsService, PbsState, PbsStateGuard, get_header, get_status,
-        register_validator, submit_block,
-    };
+    pub use cb_pbs::{PbsService, PbsState, PbsStateGuard};
     // The TreeHash derive macro requires tree_hash as import
     pub mod tree_hash {
         pub use tree_hash::*;
