+++ crates/pbs/src/lib.rs
@@ -1,13 +1,15 @@
 mod constants;
 mod error;
 mod metrics;
-mod mev_boost;
 mod routes;
 mod service;
 mod state;
 mod utils;
 
 pub use constants::*;
-pub use mev_boost::*;
+pub use routes::{
+    CompoundGetHeaderResponse, CompoundSubmitBlockResponse, LightGetHeaderResponse,
+    LightSubmitBlockResponse, get_header,
+};
 pub use service::PbsService;
-pub use state::{BuilderApiState, PbsState, PbsStateGuard};
+pub use state::{PbsState, PbsStateGuard};
