--- crates/pbs/src/mev_boost/get_header.rs
@@ -1,1019 +1,0 @@
-use std::{
-    collections::HashSet,
-    sync::Arc,
-    time::{Duration, Instant},
-};
-
-use alloy::{
-    primitives::{B256, U256, aliases::B32, utils::format_ether},
-    providers::Provider,
-    rpc::types::Block,
-};
-use axum::http::{HeaderMap, HeaderValue};
-use cb_common::{
-    config::HeaderValidationMode,
-    constants::APPLICATION_BUILDER_DOMAIN,
-    pbs::{
-        EMPTY_TX_ROOT_HASH, ExecutionPayloadHeaderRef, ForkName, ForkVersionDecode, GetHeaderInfo,
-        GetHeaderParams, GetHeaderResponse, HEADER_START_TIME_UNIX_MS, HEADER_TIMEOUT_MS,
-        RelayClient, SignedBuilderBid,
-        error::{PbsError, ValidationError},
-    },
-    signature::verify_signed_message,
-    types::{BlsPublicKey, BlsPublicKeyBytes, BlsSignature, Chain},
-    utils::{
-        EncodingType, get_bid_value_from_signed_builder_bid_ssz, get_consensus_version_header,
-        get_user_agent_with_version, ms_into_slot, read_chunked_body_with_max,
-        timestamp_of_slot_start_sec, utcnow_ms,
-    },
-};
-use futures::future::join_all;
-use parking_lot::RwLock;
-use reqwest::{
-    StatusCode,
-    header::{ACCEPT, CONTENT_TYPE, USER_AGENT},
-};
-use serde::Deserialize;
-use tokio::time::sleep;
-use tracing::{Instrument, debug, error, warn};
-use tree_hash::TreeHash;
-use url::Url;
-
-use crate::{
-    constants::{
-        GET_HEADER_ENDPOINT_TAG, MAX_SIZE_GET_HEADER_RESPONSE, TIMEOUT_ERROR_CODE,
-        TIMEOUT_ERROR_CODE_STR,
-    },
-    metrics::{RELAY_HEADER_VALUE, RELAY_LAST_SLOT, RELAY_LATENCY, RELAY_STATUS_CODE},
-    mev_boost::{CompoundGetHeaderResponse, LightGetHeaderResponse},
-    state::{BuilderApiState, PbsState},
-    utils::check_gas_limit,
-};
-
-/// Info about an incoming get_header request.
-/// Sent from get_header to each send_timed_get_header call.
-#[derive(Clone)]
-struct RequestInfo {
-    /// The blockchain parameters of the get_header request (what slot it's for,
-    /// which pubkey is requesting it, etc)
-    params: GetHeaderParams,
-
-    /// Common baseline of headers to send with each request
-    headers: Arc<HeaderMap>,
-
-    /// The chain the request is for
-    chain: Chain,
-
-    /// Context for validating the header returned by the relay
-    validation: ValidationContext,
-
-    /// The accepted encoding types from the original request
-    accepted_types: HashSet<EncodingType>,
-}
-
-/// Used interally to provide info and context about a get_header request and
-/// its response
-struct GetHeaderResponseInfo {
-    /// ID of the relay the response came from
-    relay_id: Arc<String>,
-
-    /// The raw body of the response
-    response_bytes: Vec<u8>,
-
-    /// The content type the response is encoded with
-    content_type: EncodingType,
-
-    /// Which fork the response bid is for (if provided as a header, rather than
-    /// part of the body)
-    fork: Option<ForkName>,
-
-    /// The status code of the response, for logging
-    code: StatusCode,
-
-    /// The round-trip latency of the request
-    request_latency: Duration,
-}
-
-/// Context for validating the header
-#[derive(Clone)]
-struct ValidationContext {
-    /// Whether to skip signature verification
-    skip_sigverify: bool,
-
-    /// Minimum acceptable bid, in wei
-    min_bid_wei: U256,
-
-    /// The mode used for response validation
-    mode: HeaderValidationMode,
-
-    /// The parent block, if fetched
-    parent_block: Arc<RwLock<Option<Block>>>,
-}
-
-/// Implements https://ethereum.github.io/builder-specs/#/Builder/getHeader
-/// Returns 200 if at least one relay returns 200, else 204
-pub async fn get_header<S: BuilderApiState>(
-    params: GetHeaderParams,
-    req_headers: HeaderMap,
-    state: PbsState<S>,
-    accepted_types: HashSet<EncodingType>,
-) -> eyre::Result<Option<CompoundGetHeaderResponse>> {
-    let parent_block = Arc::new(RwLock::new(None));
-    let extra_validation_enabled =
-        state.config.pbs_config.header_validation_mode == HeaderValidationMode::Extra;
-    if extra_validation_enabled && let Some(rpc_url) = state.pbs_config().rpc_url.clone() {
-        tokio::spawn(
-            fetch_parent_block(rpc_url, params.parent_hash, parent_block.clone()).in_current_span(),
-        );
-    }
-
-    let ms_into_slot = ms_into_slot(params.slot, state.config.chain);
-    let (pbs_config, relays, maybe_mux_id) = state.mux_config_and_relays(&params.pubkey);
-
-    if let Some(mux_id) = maybe_mux_id {
-        debug!(mux_id, relays = relays.len(), pubkey = %params.pubkey, "using mux config");
-    } else {
-        debug!(relays = relays.len(), pubkey = %params.pubkey, "using default config");
-    }
-
-    let max_timeout_ms = pbs_config
-        .timeout_get_header_ms
-        .min(pbs_config.late_in_slot_time_ms.saturating_sub(ms_into_slot));
-
-    if max_timeout_ms == 0 {
-        warn!(
-            ms_into_slot,
-            threshold = pbs_config.late_in_slot_time_ms,
-            "late in slot, skipping relay requests"
-        );
-
-        return Ok(None);
-    }
-
-    // Use the minimum of the time left and the user provided timeout header
-    let max_timeout_ms = req_headers
-        .get(HEADER_TIMEOUT_MS)
-        .map(|header| match header.to_str().ok().and_then(|v| v.parse::<u64>().ok()) {
-            None | Some(0) => {
-                // Header can't be stringified, or parsed, or it's set to 0
-                warn!(?header, "invalid user-supplied timeout header, using {max_timeout_ms}ms");
-                max_timeout_ms
-            }
-            Some(user_timeout) => user_timeout.min(max_timeout_ms),
-        })
-        .unwrap_or(max_timeout_ms);
-
-    // prepare headers, except for start time which is set in `send_one_get_header`
-    let mut send_headers = HeaderMap::new();
-    send_headers.insert(USER_AGENT, get_user_agent_with_version(&req_headers)?);
-
-    // Create the Accept headers for requests
-    let mode = state.pbs_config().header_validation_mode;
-    let accept_types = match mode {
-        HeaderValidationMode::None => {
-            // No validation mode, so only request what the user wants because the response
-            // will be forwarded directly
-            accepted_types.iter().map(|t| t.content_type()).collect::<Vec<&str>>().join(",")
-        }
-        _ => {
-            // We're unpacking the body, so request both types since we can handle both
-            [EncodingType::Ssz.content_type(), EncodingType::Json.content_type()].join(",")
-        }
-    };
-    send_headers.insert(
-        ACCEPT,
-        HeaderValue::from_str(&accept_types)
-            .map_err(|e| PbsError::GeneralRequest(format!("invalid accept header value: {e}")))?,
-    );
-
-    // Send requests to all relays concurrently
-    let slot = params.slot as i64;
-    let request_info = Arc::new(RequestInfo {
-        params,
-        headers: Arc::new(send_headers),
-        chain: state.config.chain,
-        validation: ValidationContext {
-            skip_sigverify: state.pbs_config().skip_sigverify,
-            min_bid_wei: state.pbs_config().min_bid_wei,
-            mode,
-            parent_block,
-        },
-        accepted_types,
-    });
-    let mut handles = Vec::with_capacity(relays.len());
-    for relay in relays.iter() {
-        handles.push(
-            send_timed_get_header(
-                request_info.clone(),
-                relay.clone(),
-                ms_into_slot,
-                max_timeout_ms,
-            )
-            .in_current_span(),
-        );
-    }
-
-    let results = join_all(handles).await;
-    let mut relay_bids = Vec::with_capacity(relays.len());
-    for (i, res) in results.into_iter().enumerate() {
-        let relay_id = relays[i].id.as_str();
-
-        match res {
-            Ok(Some(res)) => {
-                let value = match &res {
-                    CompoundGetHeaderResponse::Full(full) => *full.value(),
-                    CompoundGetHeaderResponse::Light(light) => light.value,
-                };
-                RELAY_LAST_SLOT.with_label_values(&[relay_id]).set(slot);
-                let value_gwei = (value / U256::from(1_000_000_000)).try_into().unwrap_or_default();
-                RELAY_HEADER_VALUE.with_label_values(&[relay_id]).set(value_gwei);
-
-                relay_bids.push(res)
-            }
-            Ok(_) => {}
-            Err(err) if err.is_timeout() => error!(err = "Timed Out", relay_id),
-            Err(err) => error!(%err, relay_id),
-        }
-    }
-
-    let max_bid = relay_bids.into_iter().max_by_key(|bid| match bid {
-        CompoundGetHeaderResponse::Full(full) => *full.value(),
-        CompoundGetHeaderResponse::Light(light) => light.value,
-    });
-
-    Ok(max_bid)
-}
-
-/// Fetch the parent block from the RPC URL for extra validation of the header.
-/// Extra validation will be skipped if:
-/// - relay returns header before parent block is fetched
-/// - parent block is not found, eg because of a RPC delay
-async fn fetch_parent_block(
-    rpc_url: Url,
-    parent_hash: B256,
-    parent_block: Arc<RwLock<Option<Block>>>,
-) {
-    let provider = alloy::providers::ProviderBuilder::new().connect_http(rpc_url).to_owned();
-
-    debug!(%parent_hash, "fetching parent block");
-
-    match provider.get_block_by_hash(parent_hash).await {
-        Ok(maybe_block) => {
-            debug!(block_found = maybe_block.is_some(), "fetched parent block");
-            let mut guard = parent_block.write();
-            *guard = maybe_block;
-        }
-        Err(err) => {
-            error!(%err, "fetch failed");
-        }
-    }
-}
-
-async fn send_timed_get_header(
-    request_info: Arc<RequestInfo>,
-    relay: RelayClient,
-    ms_into_slot: u64,
-    mut timeout_left_ms: u64,
-) -> Result<Option<CompoundGetHeaderResponse>, PbsError> {
-    let params = &request_info.params;
-    let url = Arc::new(relay.get_header_url(params.slot, &params.parent_hash, &params.pubkey)?);
-
-    if relay.config.enable_timing_games {
-        if let Some(target_ms) = relay.config.target_first_request_ms {
-            // sleep until target time in slot
-
-            let delay = target_ms.saturating_sub(ms_into_slot);
-            if delay > 0 {
-                debug!(
-                    relay_id = relay.id.as_ref(),
-                    target_ms, ms_into_slot, "TG: waiting to send first header request"
-                );
-                timeout_left_ms = timeout_left_ms.saturating_sub(delay);
-                sleep(Duration::from_millis(delay)).await;
-            } else {
-                debug!(
-                    relay_id = relay.id.as_ref(),
-                    target_ms, ms_into_slot, "TG: request already late enough in slot"
-                );
-            }
-        }
-
-        if let Some(send_freq_ms) = relay.config.frequency_get_header_ms {
-            let mut handles = Vec::new();
-
-            debug!(
-                relay_id = relay.id.as_ref(),
-                send_freq_ms, timeout_left_ms, "TG: sending multiple header requests"
-            );
-
-            loop {
-                handles.push(tokio::spawn(
-                    send_one_get_header(
-                        request_info.clone(),
-                        relay.clone(),
-                        url.clone(),
-                        timeout_left_ms,
-                    )
-                    .in_current_span(),
-                ));
-
-                if timeout_left_ms > send_freq_ms {
-                    // enough time for one more
-                    timeout_left_ms = timeout_left_ms.saturating_sub(send_freq_ms);
-                    sleep(Duration::from_millis(send_freq_ms)).await;
-                } else {
-                    break;
-                }
-            }
-
-            let results = join_all(handles).await;
-            let mut n_headers = 0;
-
-            if let Some((_, maybe_header)) = results
-                .into_iter()
-                .filter_map(|res| {
-                    // ignore join error and timeouts, log other errors
-                    res.ok().and_then(|inner_res| match inner_res {
-                        Ok(maybe_header) => {
-                            if maybe_header.1.is_some() {
-                                n_headers += 1;
-                                Some(maybe_header)
-                            } else {
-                                // filter out 204 responses that are returned if the request
-                                // is after the relay cutoff
-                                None
-                            }
-                        }
-                        Err(err) if err.is_timeout() => None,
-                        Err(err) => {
-                            error!(relay_id = relay.id.as_ref(),%err, "TG: error sending header request");
-                            None
-                        }
-                    })
-                })
-                .max_by_key(|(start_time, _)| *start_time)
-            {
-                debug!(relay_id = relay.id.as_ref(), n_headers, "TG: received headers from relay");
-                return Ok(maybe_header);
-            } else {
-                // all requests failed
-                warn!(relay_id = relay.id.as_ref(), "TG: no headers received");
-
-                return Err(PbsError::RelayResponse {
-                    error_msg: "no headers received".to_string(),
-                    code: TIMEOUT_ERROR_CODE,
-                });
-            }
-        }
-    }
-
-    // if no timing games or no repeated send, just send one request
-    send_one_get_header(request_info, relay, url, timeout_left_ms)
-        .await
-        .map(|(_, maybe_header)| maybe_header)
-}
-
-/// Handles requesting a header from a relay, decoding, and validation.
-/// Used by send_timed_get_header to handle each individual request.
-async fn send_one_get_header(
-    request_info: Arc<RequestInfo>,
-    relay: RelayClient,
-    url: Arc<Url>,
-    timeout_left_ms: u64,
-) -> Result<(u64, Option<CompoundGetHeaderResponse>), PbsError> {
-    match request_info.validation.mode {
-        HeaderValidationMode::None => {
-            // Minimal processing: extract fork and value, forward response bytes directly.
-            // Expensive crypto/structural validation is skipped (sigverify, parent hash,
-            // timestamp), but the min_bid check is applied.
-            let (start_request_time, get_header_response) = send_get_header_light(
-                &relay,
-                url,
-                timeout_left_ms,
-                (*request_info.headers).clone(), /* Create a copy of the HeaderMap because the
-                                                  * impl
-                                                  * will
-                                                  * modify it */
-            )
-            .await?;
-            match get_header_response {
-                None => Ok((start_request_time, None)),
-                Some(res) => {
-                    let min_bid = request_info.validation.min_bid_wei;
-                    if res.value < min_bid {
-                        return Err(PbsError::Validation(ValidationError::BidTooLow {
-                            min: min_bid,
-                            got: res.value,
-                        }));
-                    }
-
-                    // Make sure the response is encoded in one of the accepted
-                    // types since we're passing the raw response directly to the client
-                    if !request_info.accepted_types.contains(&res.encoding_type) {
-                        return Err(PbsError::RelayResponse {
-                            error_msg: format!(
-                                "relay returned unsupported encoding type for get_header in no-validation mode: {:?}",
-                                res.encoding_type
-                            ),
-                            code: 406, // Not Acceptable
-                        });
-                    }
-                    Ok((start_request_time, Some(CompoundGetHeaderResponse::Light(res))))
-                }
-            }
-        }
-        _ => {
-            // Full processing: decode full response and validate
-            let (start_request_time, get_header_response) = send_get_header_full(
-                &relay,
-                url,
-                timeout_left_ms,
-                (*request_info.headers).clone(), /* Create a copy of the HeaderMap because the
-                                                  * impl
-                                                  * will
-                                                  * modify it */
-            )
-            .await?;
-            let get_header_response = match get_header_response {
-                None => {
-                    // Break if there's no header
-                    return Ok((start_request_time, None));
-                }
-                Some(res) => res,
-            };
-
-            // Extract the basic header data needed for validation
-            let header_data = match &get_header_response.data.message.header() {
-                ExecutionPayloadHeaderRef::Bellatrix(_) |
-                ExecutionPayloadHeaderRef::Capella(_) |
-                ExecutionPayloadHeaderRef::Deneb(_) |
-                ExecutionPayloadHeaderRef::Gloas(_) => {
-                    Err(PbsError::Validation(ValidationError::UnsupportedFork))
-                }
-                ExecutionPayloadHeaderRef::Electra(res) => Ok(HeaderData {
-                    block_hash: res.block_hash.0,
-                    parent_hash: res.parent_hash.0,
-                    tx_root: res.transactions_root,
-                    value: *get_header_response.value(),
-                    timestamp: res.timestamp,
-                }),
-                ExecutionPayloadHeaderRef::Fulu(res) => Ok(HeaderData {
-                    block_hash: res.block_hash.0,
-                    parent_hash: res.parent_hash.0,
-                    tx_root: res.transactions_root,
-                    value: *get_header_response.value(),
-                    timestamp: res.timestamp,
-                }),
-            }?;
-
-            // Validate the header
-            let chain = request_info.chain;
-            let params = &request_info.params;
-            let validation = &request_info.validation;
-            validate_header_data(
-                &header_data,
-                chain,
-                params.parent_hash,
-                validation.min_bid_wei,
-                params.slot,
-            )?;
-
-            // Validate the relay signature
-            if !validation.skip_sigverify {
-                validate_signature(
-                    chain,
-                    relay.pubkey(),
-                    get_header_response.data.message.pubkey(),
-                    &get_header_response.data.message,
-                    &get_header_response.data.signature,
-                )?;
-            }
-
-            // Validate the parent block if enabled
-            if validation.mode == HeaderValidationMode::Extra {
-                let parent_block = validation.parent_block.read();
-                if let Some(parent_block) = parent_block.as_ref() {
-                    extra_validation(parent_block, &get_header_response)?;
-                } else {
-                    warn!(
-                        relay_id = relay.id.as_ref(),
-                        "parent block not found, skipping extra validation"
-                    );
-                }
-            }
-
-            Ok((
-                start_request_time,
-                Some(CompoundGetHeaderResponse::Full(Box::new(get_header_response))),
-            ))
-        }
-    }
-}
-
-/// Send and decode a full get_header response, will all of the fields.
-async fn send_get_header_full(
-    relay: &RelayClient,
-    url: Arc<Url>,
-    timeout_left_ms: u64,
-    headers: HeaderMap,
-) -> Result<(u64, Option<GetHeaderResponse>), PbsError> {
-    // Send the request
-    let (start_request_time, info) =
-        send_get_header_impl(relay, url, timeout_left_ms, headers).await?;
-    let info = match info {
-        Some(info) => info,
-        None => {
-            return Ok((start_request_time, None));
-        }
-    };
-
-    // Decode the response
-    let get_header_response = match info.content_type {
-        EncodingType::Json => decode_json_payload(&info.response_bytes)?,
-        EncodingType::Ssz => {
-            let fork = info.fork.ok_or(PbsError::RelayResponse {
-                error_msg: "relay did not provide consensus version header for ssz payload"
-                    .to_string(),
-                code: info.code.as_u16(),
-            })?;
-            decode_ssz_payload(&info.response_bytes, fork)?
-        }
-    };
-
-    // Log and return
-    debug!(
-        relay_id = info.relay_id.as_ref(),
-        header_size_bytes = info.response_bytes.len(),
-        latency = ?info.request_latency,
-        version =? get_header_response.version,
-        value_eth = format_ether(*get_header_response.value()),
-        block_hash = %get_header_response.block_hash(),
-        content_type = ?info.content_type,
-        "received new header"
-    );
-    Ok((start_request_time, Some(get_header_response)))
-}
-
-/// Send a get_header request and decode only the fork and bid value from the
-/// response, leaving the raw bytes intact for direct forwarding to the caller.
-/// Used in `HeaderValidationMode::None` where expensive crypto/structural
-/// checks are skipped.
-async fn send_get_header_light(
-    relay: &RelayClient,
-    url: Arc<Url>,
-    timeout_left_ms: u64,
-    headers: HeaderMap,
-) -> Result<(u64, Option<LightGetHeaderResponse>), PbsError> {
-    // Send the request
-    let (start_request_time, info) =
-        send_get_header_impl(relay, url, timeout_left_ms, headers).await?;
-    let info = match info {
-        Some(info) => info,
-        None => {
-            return Ok((start_request_time, None));
-        }
-    };
-
-    // Decode the value / fork from the response
-    let (fork, value) = match info.content_type {
-        EncodingType::Json => get_light_info_from_json(&info.response_bytes)?,
-        EncodingType::Ssz => {
-            let fork = info.fork.ok_or(PbsError::RelayResponse {
-                error_msg: "relay did not provide consensus version header for ssz payload"
-                    .to_string(),
-                code: info.code.as_u16(),
-            })?;
-            (fork, get_bid_value_from_signed_builder_bid_ssz(&info.response_bytes, fork)?)
-        }
-    };
-
-    // Log and return
-    debug!(
-        relay_id = info.relay_id.as_ref(),
-        header_size_bytes = info.response_bytes.len(),
-        latency = ?info.request_latency,
-        version =? fork,
-        value_eth = format_ether(value),
-        content_type = ?info.content_type,
-        "received new header (light processing)"
-    );
-    Ok((
-        start_request_time,
-        Some(LightGetHeaderResponse {
-            version: fork,
-            value,
-            raw_bytes: info.response_bytes,
-            encoding_type: info.content_type,
-        }),
-    ))
-}
-
-/// Sends a get_header request to a relay, returning the response, the time the
-/// request was started, and the encoding type of the response (if any).
-/// Used by send_one_get_header to perform the actual request submission.
-async fn send_get_header_impl(
-    relay: &RelayClient,
-    url: Arc<Url>,
-    timeout_left_ms: u64,
-    mut headers: HeaderMap,
-) -> Result<(u64, Option<GetHeaderResponseInfo>), PbsError> {
-    // the timestamp in the header is the consensus block time which is fixed,
-    // use the beginning of the request as proxy to make sure we use only the
-    // last one received
-    let start_request = Instant::now();
-    let start_request_time = utcnow_ms();
-    headers.insert(HEADER_START_TIME_UNIX_MS, HeaderValue::from(start_request_time));
-
-    // The timeout header indicating how long a relay has to respond, so they can
-    // minimize timing games without losing the bid
-    headers.insert(HEADER_TIMEOUT_MS, HeaderValue::from(timeout_left_ms));
-
-    let res = match relay
-        .client
-        .get(url.as_ref().clone())
-        .timeout(Duration::from_millis(timeout_left_ms))
-        .headers(headers)
-        .send()
-        .await
-    {
-        Ok(res) => res,
-        Err(err) => {
-            RELAY_STATUS_CODE
-                .with_label_values(&[TIMEOUT_ERROR_CODE_STR, GET_HEADER_ENDPOINT_TAG, &relay.id])
-                .inc();
-            return Err(err.into());
-        }
-    };
-
-    // Log the response code and latency
-    let code = res.status();
-    let request_latency = start_request.elapsed();
-    RELAY_LATENCY
-        .with_label_values(&[GET_HEADER_ENDPOINT_TAG, &relay.id])
-        .observe(request_latency.as_secs_f64());
-    RELAY_STATUS_CODE.with_label_values(&[code.as_str(), GET_HEADER_ENDPOINT_TAG, &relay.id]).inc();
-
-    // According to the spec, OK is the only allowed success code so this can break
-    // early
-    if code != StatusCode::OK {
-        if code == StatusCode::NO_CONTENT {
-            let response_bytes =
-                read_chunked_body_with_max(res, MAX_SIZE_GET_HEADER_RESPONSE).await?;
-            debug!(
-                relay_id = relay.id.as_ref(),
-                ?code,
-                latency = ?request_latency,
-                response = ?response_bytes,
-                "no header from relay"
-            );
-            return Ok((start_request_time, None));
-        } else {
-            return Err(PbsError::RelayResponse {
-                error_msg: format!("unexpected status code from relay: {code}"),
-                code: code.as_u16(),
-            });
-        }
-    }
-
-    // Get the content type
-    let content_type = match res.headers().get(CONTENT_TYPE) {
-        None => {
-            // Assume a missing content type means JSON; shouldn't happen in practice with
-            // any respectable HTTP server but just in case
-            EncodingType::Json
-        }
-        Some(header_value) => match header_value.to_str().map_err(|e| PbsError::RelayResponse {
-            error_msg: format!("cannot decode content-type header: {e}").to_string(),
-            code: (code.as_u16()),
-        })? {
-            header_str if header_str.eq_ignore_ascii_case(&EncodingType::Ssz.to_string()) => {
-                EncodingType::Ssz
-            }
-            header_str if header_str.eq_ignore_ascii_case(&EncodingType::Json.to_string()) => {
-                EncodingType::Json
-            }
-            header_str => {
-                return Err(PbsError::RelayResponse {
-                    error_msg: format!("unsupported content type: {header_str}"),
-                    code: code.as_u16(),
-                })
-            }
-        },
-    };
-
-    // Decode the body
-    let fork = get_consensus_version_header(res.headers());
-    let response_bytes = read_chunked_body_with_max(res, MAX_SIZE_GET_HEADER_RESPONSE).await?;
-    Ok((
-        start_request_time,
-        Some(GetHeaderResponseInfo {
-            relay_id: relay.id.clone(),
-            response_bytes,
-            content_type,
-            fork,
-            code,
-            request_latency,
-        }),
-    ))
-}
-
-/// Decode a JSON-encoded get_header response
-fn decode_json_payload(response_bytes: &[u8]) -> Result<GetHeaderResponse, PbsError> {
-    match serde_json::from_slice::<GetHeaderResponse>(response_bytes) {
-        Ok(parsed) => Ok(parsed),
-        Err(err) => Err(PbsError::JsonDecode {
-            err,
-            raw: String::from_utf8_lossy(response_bytes).into_owned(),
-        }),
-    }
-}
-
-/// Get the value of a builder bid and the fork name from a get_header JSON
-/// response (used for light-level processing)
-fn get_light_info_from_json(response_bytes: &[u8]) -> Result<(ForkName, U256), PbsError> {
-    #[derive(Deserialize)]
-    struct LightBuilderBid {
-        #[serde(with = "serde_utils::quoted_u256")]
-        pub value: U256,
-    }
-
-    #[derive(Deserialize)]
-    struct LightSignedBuilderBid {
-        pub message: LightBuilderBid,
-    }
-
-    #[derive(Deserialize)]
-    struct LightHeaderResponse {
-        version: ForkName,
-        data: LightSignedBuilderBid,
-    }
-
-    match serde_json::from_slice::<LightHeaderResponse>(response_bytes) {
-        Ok(parsed) => Ok((parsed.version, parsed.data.message.value)),
-        Err(err) => Err(PbsError::JsonDecode {
-            err,
-            raw: String::from_utf8_lossy(response_bytes).into_owned(),
-        }),
-    }
-}
-
-/// Decode an SSZ-encoded get_header response
-fn decode_ssz_payload(
-    response_bytes: &[u8],
-    fork: ForkName,
-) -> Result<GetHeaderResponse, PbsError> {
-    let data = SignedBuilderBid::from_ssz_bytes_by_fork(response_bytes, fork).map_err(|e| {
-        PbsError::RelayResponse {
-            error_msg: (format!("error decoding relay payload: {e:?}")).to_string(),
-            code: 200,
-        }
-    })?;
-    Ok(GetHeaderResponse { version: fork, data, metadata: Default::default() })
-}
-
-struct HeaderData {
-    block_hash: B256,
-    parent_hash: B256,
-    tx_root: B256,
-    value: U256,
-    timestamp: u64,
-}
-
-fn validate_header_data(
-    header_data: &HeaderData,
-    chain: Chain,
-    expected_parent_hash: B256,
-    minimum_bid_wei: U256,
-    slot: u64,
-) -> Result<(), ValidationError> {
-    if header_data.block_hash == B256::ZERO {
-        return Err(ValidationError::EmptyBlockhash);
-    }
-
-    if expected_parent_hash != header_data.parent_hash {
-        return Err(ValidationError::ParentHashMismatch {
-            expected: expected_parent_hash,
-            got: header_data.parent_hash,
-        });
-    }
-
-    if header_data.tx_root == EMPTY_TX_ROOT_HASH {
-        return Err(ValidationError::EmptyTxRoot);
-    }
-
-    if header_data.value < minimum_bid_wei {
-        return Err(ValidationError::BidTooLow { min: minimum_bid_wei, got: header_data.value });
-    }
-
-    let expected_timestamp = timestamp_of_slot_start_sec(slot, chain);
-    if expected_timestamp != header_data.timestamp {
-        return Err(ValidationError::TimestampMismatch {
-            expected: expected_timestamp,
-            got: header_data.timestamp,
-        });
-    }
-
-    Ok(())
-}
-
-fn validate_signature<T: TreeHash>(
-    chain: Chain,
-    expected_relay_pubkey: &BlsPublicKey,
-    received_relay_pubkey: &BlsPublicKeyBytes,
-    message: &T,
-    signature: &BlsSignature,
-) -> Result<(), ValidationError> {
-    if expected_relay_pubkey.serialize() != received_relay_pubkey.as_serialized() {
-        return Err(ValidationError::PubkeyMismatch {
-            expected: BlsPublicKeyBytes::from(expected_relay_pubkey),
-            got: *received_relay_pubkey,
-        });
-    }
-
-    if !verify_signed_message(
-        chain,
-        expected_relay_pubkey,
-        &message,
-        signature,
-        None,
-        &B32::from(APPLICATION_BUILDER_DOMAIN),
-    ) {
-        return Err(ValidationError::Sigverify);
-    }
-
-    Ok(())
-}
-
-fn extra_validation(
-    parent_block: &Block,
-    signed_header: &GetHeaderResponse,
-) -> Result<(), ValidationError> {
-    if signed_header.block_number() != parent_block.header.number + 1 {
-        return Err(ValidationError::BlockNumberMismatch {
-            parent: parent_block.header.number,
-            header: signed_header.block_number(),
-        });
-    }
-
-    if !check_gas_limit(signed_header.gas_limit(), parent_block.header.gas_limit) {
-        return Err(ValidationError::GasLimit {
-            parent: parent_block.header.gas_limit,
-            header: signed_header.gas_limit(),
-        });
-    };
-
-    Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-    use std::{fs, path::Path};
-
-    use alloy::primitives::{B256, U256};
-    use cb_common::{
-        pbs::*,
-        signature::sign_builder_message,
-        types::{BlsPublicKeyBytes, BlsSecretKey, BlsSignature, Chain},
-        utils::{TestRandomSeed, timestamp_of_slot_start_sec},
-    };
-    use ssz::Encode;
-
-    use super::{validate_header_data, *};
-
-    #[test]
-    fn test_validate_header() {
-        let slot = 5;
-        let parent_hash = B256::from_slice(&[1; 32]);
-        let chain = Chain::Holesky;
-        let min_bid = U256::from(10);
-
-        let mut mock_header_data = HeaderData {
-            block_hash: B256::default(),
-            parent_hash: B256::default(),
-            tx_root: EMPTY_TX_ROOT_HASH,
-            value: U256::default(),
-            timestamp: 0,
-        };
-
-        assert_eq!(
-            validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot,),
-            Err(ValidationError::EmptyBlockhash)
-        );
-
-        mock_header_data.block_hash.0[1] = 1;
-
-        assert_eq!(
-            validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot,),
-            Err(ValidationError::ParentHashMismatch {
-                expected: parent_hash,
-                got: B256::default()
-            })
-        );
-
-        mock_header_data.parent_hash = parent_hash;
-
-        assert_eq!(
-            validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot,),
-            Err(ValidationError::EmptyTxRoot)
-        );
-
-        mock_header_data.tx_root = Default::default();
-
-        assert_eq!(
-            validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot,),
-            Err(ValidationError::BidTooLow { min: min_bid, got: U256::ZERO })
-        );
-
-        mock_header_data.value = U256::from(11);
-
-        let expected = timestamp_of_slot_start_sec(slot, chain);
-        assert_eq!(
-            validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot,),
-            Err(ValidationError::TimestampMismatch { expected, got: 0 })
-        );
-
-        mock_header_data.timestamp = expected;
-
-        assert!(validate_header_data(&mock_header_data, chain, parent_hash, min_bid, slot).is_ok());
-    }
-
-    #[test]
-    fn test_validate_signature() {
-        let secret_key = BlsSecretKey::test_random();
-        let pubkey = secret_key.public_key();
-        let wrong_pubkey = BlsPublicKeyBytes::test_random();
-        let wrong_signature = BlsSignature::test_random();
-
-        let message = B256::random();
-
-        let signature = sign_builder_message(Chain::Holesky, &secret_key, &message);
-
-        assert_eq!(
-            validate_signature(Chain::Holesky, &pubkey, &wrong_pubkey, &message, &wrong_signature),
-            Err(ValidationError::PubkeyMismatch {
-                expected: BlsPublicKeyBytes::from(&pubkey),
-                got: wrong_pubkey
-            })
-        );
-
-        assert!(matches!(
-            validate_signature(
-                Chain::Holesky,
-                &pubkey,
-                &BlsPublicKeyBytes::from(&pubkey),
-                &message,
-                &wrong_signature
-            ),
-            Err(ValidationError::Sigverify)
-        ));
-
-        assert!(
-            validate_signature(
-                Chain::Holesky,
-                &pubkey,
-                &BlsPublicKeyBytes::from(&pubkey),
-                &message,
-                &signature
-            )
-            .is_ok()
-        );
-    }
-
-    #[test]
-    fn test_ssz_value_extraction() {
-        for fork_name in ForkName::list_all() {
-            match fork_name {
-                // Handle forks that didn't have builder bids yet
-                ForkName::Altair | ForkName::Base => continue,
-
-                // Handle supported forks
-                ForkName::Bellatrix |
-                ForkName::Capella |
-                ForkName::Deneb |
-                ForkName::Electra |
-                ForkName::Fulu => {}
-
-                // Skip unsupported forks
-                ForkName::Gloas => continue,
-            }
-
-            // Load get_header JSON from test data
-            let fork_name_str = fork_name.to_string().to_lowercase();
-            let path_str = format!("../../tests/data/get_header/{fork_name_str}.json");
-            let path = Path::new(path_str.as_str());
-            let json_bytes = fs::read(path).expect("file not found");
-            let decoded = decode_json_payload(&json_bytes).expect("failed to decode JSON");
-
-            // Extract the bid value from the SSZ
-            let encoded = decoded.data.as_ssz_bytes();
-            let bid_value = get_bid_value_from_signed_builder_bid_ssz(&encoded, fork_name)
-                .expect("failed to extract bid value from SSZ");
-
-            // Compare to the original value
-            println!("Testing fork: {}", fork_name);
-            println!("Original value: {}", decoded.value());
-            println!("Extracted value: {}", bid_value);
-            assert_eq!(*decoded.value(), bid_value);
-        }
-    }
-}
