+++ tests/tests/pbs_mux_refresh.rs
@@ -8,10 +8,10 @@ use cb_common::{
 };
 use cb_pbs::{PbsService, PbsState};
 use cb_tests::{
-    mock_relay::{MockRelayState, start_mock_relay_service},
-    mock_ssv_public::{PublicSsvMockState, create_mock_public_ssv_server},
+    mock_relay::{MockRelayState, start_mock_relay_service_with_listener},
+    mock_ssv_public::{PublicSsvMockState, create_mock_public_ssv_server_with_listener},
     mock_validator::MockValidator,
-    utils::{generate_mock_relay, get_pbs_config, to_pbs_config},
+    utils::{generate_mock_relay, get_free_listener, get_pbs_config, to_pbs_config},
 };
 use eyre::Result;
 use lh_types::ForkName;
@@ -39,10 +39,15 @@ async fn test_auto_refresh() -> Result<()> {
     let new_mux_pubkey = new_mux_signer.public_key();
 
     let chain = Chain::Hoodi;
-    let pbs_port = 3710;
+    let pbs_listener = get_free_listener().await;
+    let ssv_listener = get_free_listener().await;
+    let default_relay_listener = get_free_listener().await;
+    let mux_relay_listener = get_free_listener().await;
+    let pbs_port = pbs_listener.local_addr().unwrap().port();
+    let ssv_api_port = ssv_listener.local_addr().unwrap().port();
+    let default_relay_port = default_relay_listener.local_addr().unwrap().port();
+    let mux_relay_port = mux_relay_listener.local_addr().unwrap().port();
 
-    // Start the mock SSV API server
-    let ssv_api_port = pbs_port + 1;
     // Intentionally missing a trailing slash to ensure this is handled properly
     let ssv_api_url = Url::parse(&format!("http://localhost:{ssv_api_port}/api/v4"))?;
     let mock_ssv_state = PublicSsvMockState {
@@ -51,23 +56,24 @@ async fn test_auto_refresh() -> Result<()> {
         }])),
         force_timeout: Arc::new(RwLock::new(false)),
     };
-    let ssv_server_handle =
-        create_mock_public_ssv_server(ssv_api_port, Some(mock_ssv_state.clone())).await?;
+    create_mock_public_ssv_server_with_listener(ssv_listener, Some(mock_ssv_state.clone())).await?;
 
     // Start a default relay for non-mux keys
-    let default_relay_port = ssv_api_port + 1;
     let default_relay = generate_mock_relay(default_relay_port, default_pubkey.clone())?;
     let default_relay_state = Arc::new(MockRelayState::new(chain, default_signer.clone()));
-    let default_relay_task =
-        tokio::spawn(start_mock_relay_service(default_relay_state.clone(), default_relay_port));
+    let default_relay_task = tokio::spawn(start_mock_relay_service_with_listener(
+        default_relay_state.clone(),
+        default_relay_listener,
+    ));
 
     // Start a mock relay to be used by the mux
-    let mux_relay_port = default_relay_port + 1;
     let mux_relay = generate_mock_relay(mux_relay_port, default_pubkey.clone())?;
     let mux_relay_id = mux_relay.id.clone().to_string();
     let mux_relay_state = Arc::new(MockRelayState::new(chain, default_signer));
-    let mux_relay_task =
-        tokio::spawn(start_mock_relay_service(mux_relay_state.clone(), mux_relay_port));
+    let mux_relay_task = tokio::spawn(start_mock_relay_service_with_listener(
+        mux_relay_state.clone(),
+        mux_relay_listener,
+    ));
 
     // Create the registry mux
     let loader = MuxKeysLoader::Registry {
@@ -100,7 +106,7 @@ async fn test_auto_refresh() -> Result<()> {
 
     // Run PBS service
     let state = PbsState::new(config, PathBuf::new());
-    let pbs_server = tokio::spawn(PbsService::run::<()>(state));
+    let pbs_server = tokio::spawn(PbsService::run_with_listener(state, pbs_listener));
     info!("Started PBS server with pubkey {default_pubkey}");
 
     // Wait for the server to start
@@ -177,7 +183,6 @@ async fn test_auto_refresh() -> Result<()> {
 
     // Shut down the server handles
     pbs_server.abort();
-    ssv_server_handle.abort();
     default_relay_task.abort();
     mux_relay_task.abort();
